requirements.txt
asgiref==3.9.1
asyncpg==0.30.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
greenlet==3.2.4
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
